some useful changes regarding space and time adaptivity (state) - fixed hopper issue with escaping particles (no-grid, regular, adaptive)

+++ delta_peano/dem/Vertex.cpp
@@ -557,7 +557,7 @@ void dem::Vertex::setNumberOfParticlesInUnrefinedVertex(double number)
   _vertexData.setNumberOfParticlesInUnrefinedVertex(number);
 }
 
-int dem::Vertex::getNumberOfParticlesInUnrefinedVertex()
+double dem::Vertex::getNumberOfParticlesInUnrefinedVertex()
 {
   return _vertexData.getNumberOfParticlesInUnrefinedVertex();
 }
